+++ packages/c6o-system/src/mixins/api/index.ts
@@ -1,4 +1,4 @@
-export * from './createAppPost'
+export * from './postApp'
 export * from './choices'
 export * from './logger'
 export * from './npm'