+++ packages/mattermost/CHANGELOG.md
@@ -3,6 +3,33 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.0.7](https://github.com/c6o/provisioners/compare/v0.0.6...v0.0.7) (2020-09-03)
+
+
+### ♻️ Chores
+
+* Bumped gitHead ([15d30cf](https://github.com/c6o/provisioners/commit/15d30cf8f5386a58e2873cf2dd97fdc55f8f7cd2))
+* Bumped versions ([1da2565](https://github.com/c6o/provisioners/commit/1da25659e5cbe7989a20537e62f2cc730005a699))
+* Bumped versions ([43a02cb](https://github.com/c6o/provisioners/commit/43a02cbde47a491bd5d318fe896d9922cdc71ba4))
+
+
+
+
+
+## [0.0.6](https://github.com/c6o/provisioners/compare/v0.0.5...v0.0.6) (2020-08-23)
+
+
+### ♻️ Chores
+
+* Bumped references ([798c6a3](https://github.com/c6o/provisioners/commit/798c6a3f7c826d04f2327a5cfae535f2dd3d04e8))
+* Removed extraneous debugger statement ([2981418](https://github.com/c6o/provisioners/commit/298141896a221c585856f2d11541ced346556eb8))
+* Spelling issues ([f302134](https://github.com/c6o/provisioners/commit/f3021348e25e9cb1f76cce1694d2cdba236b491d))
+* Updated dependencies ([18ff8de](https://github.com/c6o/provisioners/commit/18ff8de36b5a0c9b1f343842724167fdc60bf62d))
+
+
+
+
+
 ## [0.0.5](https://github.com/c6o/provisioners/compare/v0.0.4...v0.0.5) (2020-08-13)